Adds fonts
