+++ HuntGuide/src/main/java/me/androidbox/huntguide/presentation/screens/HuntGuideScreen.kt
@@ -21,7 +21,7 @@ fun HuntGuideScreen(
         modifier = Modifier.background(color = Color.Black),
         items = huntGuideItems,
         content = { huntGuideItem ->
-            HuntGuideItemPage(
+            HuntGuidePageItem(
                 title = huntGuideItem.title,
                 subTitle = huntGuideItem.subTitle,
                 body = huntGuideItem.body,
